// src/types/global.d.ts
export interface Pagination {
  total: number;
  count: number;
  page: number;
  totalPages: number;
  hasMore: boolean;
}